+++ backend/packages/spend/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @cat-factory/spend
 
+## 0.15.29
+
+### Patch Changes
+
+- Updated dependencies [1025674]
+- Updated dependencies [e5f7eb0]
+  - @cat-factory/contracts@0.263.0
+  - @cat-factory/kernel@0.262.1
+
+## 0.15.28
+
+### Patch Changes
+
+- Updated dependencies [8cbd518]
+- Updated dependencies [8cbd518]
+- Updated dependencies [7a2730a]
+  - @cat-factory/contracts@0.262.0
+  - @cat-factory/kernel@0.262.0
+
 ## 0.15.27
 
 ### Patch Changes